WinFormDemo single thread starting point
This is only a starting point. I left some unreachable code
and did not fully clean things up. The purpose was to determine
if the multithreaded nature of the demo was interfering with the
Raman 532 ER.

Now that this program can get spectra single-threaded (working with
XS), I see that the 532 still fails. So I should continue my
investigation elsewhere... Maybe device initialization or something
else in the core driver is interfering with the device's normal
operation.